manifest-c9s: add SIGs GPG keys
Add the GPG keys for the various related SIGs to the base image. It's
not strictly needed for the base package set, but is needed for the node
image build and extensions build. We've previously been adding those
packages there, but it causes a lot of pain in the associated logic and
it just doesn't feel worth it anymore.