get-ocp-repo.sh: drastically simplify
Now that (1) we've reworked the layered node image build to only enable
the repos it needs, and (2) we've simplified the CentOS Stream GPG keys,
we can delete all of the complex logic in this repo. It basically just
boils down to curl'ing down all the repo files we may need to build the
various artifacts that use this script.